fix(connectors): clippy lint fixes, Debug redaction improvements, and input validation
- Rename `_id` bindings in CredentialId Debug impls to avoid clippy
  warnings about underscore-prefixed names being read (brave-search,
  deepgram, elevenlabs, exa, mistral, tavily)
- Collapse nested Err or-patterns into idiomatic `Err(A | B)` form
  (amplitude, posthog)
- Switch Debug impls to finish_non_exhaustive() where struct has
  private fields (discord, llm-router)
- Narrow google-places error code params from u32 to u16
- Remove redundant let-binding in llm-router retry loop
- Mark discord api_client() as const fn
- Add amplitude config_rejects_object_secret_key test

+++ connectors/google-places/src/types.rs
@@ -167,19 +167,19 @@ fn default_place_details_field_mask() -> String {
     DEFAULT_PLACE_DETAILS_FIELD_MASK.to_string()
 }
 
-fn invalid_request(code: u32, message: impl Into<String>) -> FcpError {
+fn invalid_request(code: u16, message: impl Into<String>) -> FcpError {
     FcpError::InvalidRequest {
         code,
         message: message.into(),
     }
 }
 
-fn parse_value<T: DeserializeOwned>(value: Value, code: u32, context: &str) -> FcpResult<T> {
+fn parse_value<T: DeserializeOwned>(value: Value, code: u16, context: &str) -> FcpResult<T> {
     serde_json::from_value(value)
         .map_err(|error| invalid_request(code, format!("Invalid {context}: {error}")))
 }
 
-fn validate_non_empty_mask(mask: &str, field: &str, code: u32) -> FcpResult<()> {
+fn validate_non_empty_mask(mask: &str, field: &str, code: u16) -> FcpResult<()> {
     if mask.trim().is_empty() {
         return Err(invalid_request(code, format!("{field} must not be empty")));
     }
